refactor: opponent, dto, converter, exception 패키지 위치 수정
