+++ src/main/java/baseball/console/converter/ThreeNumberConverter.java
@@ -1,10 +1,10 @@
-package baseball.converter;
+package baseball.console.converter;
 
-import baseball.exception.DuplicateNumberException;
-import baseball.exception.EachInputNumberOutOfRangeException;
-import baseball.exception.InputNumberOutOfRangeException;
-import baseball.exception.IllegalNumberInputException;
-import baseball.exception.InputNumberFormatException;
+import baseball.console.exception.DuplicateNumberException;
+import baseball.console.exception.EachInputNumberOutOfRangeException;
+import baseball.console.exception.InputNumberOutOfRangeException;
+import baseball.console.exception.IllegalNumberInputException;
+import baseball.console.exception.InputNumberFormatException;
 
 import java.util.HashSet;
 import java.util.List;
