+++ src/main/java/baseball/console/exception/EachInputNumberOutOfRangeException.java
@@ -1,4 +1,4 @@
-package baseball.exception;
+package baseball.console.exception;
 
 public class EachInputNumberOutOfRangeException extends IllegalNumberInputException {
