+++ src/main/java/baseball/console/exception/InputNumberFormatException.java
@@ -1,4 +1,4 @@
-package baseball.exception;
+package baseball.console.exception;
 
 public class InputNumberFormatException extends IllegalNumberInputException {
