+++ src/main/java/baseball/console/exception/InputNumberOutOfRangeException.java
@@ -1,4 +1,4 @@
-package baseball.exception;
+package baseball.console.exception;
 
 public class InputNumberOutOfRangeException extends IllegalNumberInputException {
