+++ src/main/java/baseball/core/dto/BallStrikeDto.java
@@ -1,4 +1,4 @@
-package baseball.dto;
+package baseball.core.dto;
 
 public class BallStrikeDto {
