+++ src/main/java/baseball/printer/ResultMessageGenerator.java
@@ -1,6 +1,6 @@
 package baseball.printer;
 
-import baseball.dto.BallStrikeDto;
+import baseball.core.dto.BallStrikeDto;
 
 import java.util.List;
