+++ src/test/java/baseball/opponent/ComputerOpponentTest.java
@@ -1,6 +1,7 @@
 package baseball.opponent;
 
-import baseball.dto.BallStrikeDto;
+import baseball.core.opponent.ComputerOpponent;
+import baseball.core.dto.BallStrikeDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -8,7 +9,6 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class ComputerOpponentTest {
